+++ OsvvmSettingsRequired.tcl
@@ -56,7 +56,7 @@
 
 namespace eval ::osvvm {
 
-  variable OsvvmVersion 2024.0X-Dev
+  variable OsvvmVersion 2024.05
 
   if {![info exists OsvvmVersionCompatibility]} {
     variable OsvvmVersionCompatibility $OsvvmVersion