+++ ReportAlert2Html.tcl
@@ -19,6 +19,7 @@
 #
 #  Revision History:
 #    Date      Version    Description
+#    05/2024   2024.05    Minor updates during Simulate2Html refactoring
 #    04/2024   2024.04    Updated report formatting
 #    02/2022   2022.02    Updated YAML file handling
 #    10/2021   Initial    Initial Revision