+++ VendorScripts_NVC.tcl
@@ -19,6 +19,7 @@
 # 
 #  Revision History:
 #    Date      Version    Description
+#     5/2024   2024.05    Added ToolVersion variable 
 #     1/2023   2023.01    Added options for CoSim 
 #    10/2022              Initial Version based on VendorScripts_GHDL.tcl
 #